--- src/Controller/BookController.php
@@ -1,109 +1,0 @@
-<?php
-
-namespace App\Controller;
-
-use App\Entity\Book;
-use App\Form\BookType;
-use App\Repository\BookRepository;
-use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Attribute\Route;
-use Symfony\Component\Serializer\SerializerInterface;
-use Symfony\Component\String\Slugger\SluggerInterface;
-
-#[Route('/book')]
-final class BookController extends AbstractController
-{
-    #[Route('/api', name: 'app_book_api')]
-    public function api(BookRepository $bookRepo, SerializerInterface $serializer): Response
-    {
-        return JsonResponse::fromJsonString($serializer->serialize($bookRepo->findAll(), 'json', ['groups' => ['book:read']]));
-    }
-
-    // http://localhost:8000/book/api
-
-    #[Route('/new', name: 'app_book_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager, SluggerInterface $slugger): Response
-    {
-        if (!$this->isGranted('ROLE_ADMIN')) return $this->redirectToRoute('app_home');
-        $book = new Book();
-        $form = $this->createForm(BookType::class, $book);
-        $form->handleRequest($request);
-    
-        if ($form->isSubmitted() && $form->isValid()) {
-            $imageFile = $form->get('cover')->getData();
-            if ($imageFile) {
-                $originalFilename = pathinfo($imageFile->getClientOriginalName(), PATHINFO_FILENAME);
-                $safeFilename = $slugger->slug($originalFilename);
-                $newFilename = $safeFilename . '-' . uniqid() . '.' . $imageFile->guessExtension();
-                $imageFile->move(
-                    $this->getParameter('images_directory'),
-                    $newFilename
-                );
-                $book->setCover($newFilename);
-            }
-    
-            $entityManager->persist($book);
-            $entityManager->flush();
-    
-            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
-        }
-    
-        return $this->render('book/new.html.twig', [
-            'book' => $book,
-            'form' => $form,
-        ]);
-    }
-
-    #[Route('/{id}', name: 'app_book_delete', methods: ['POST'])]
-    public function delete(Request $request, Book $book, EntityManagerInterface $entityManager): Response
-    {
-        if ($this->isCsrfTokenValid('delete' . $book->getId(), $request->getPayload()->getString('_token'))) {
-            $entityManager->remove($book);
-            $entityManager->flush();
-        }
-        return $this->redirectToRoute('app_dashboard', [], Response::HTTP_SEE_OTHER);
-    }
-
-    #[Route('/{id}', name: 'app_book', methods: ['GET'])]
-    public function book(Book $book): Response
-    {
-        return $this->render('main/book.html.twig', [
-            'book' => $book
-        ]);
-    }
-
-    #[Route('/{id}/edit', name: 'app_book_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Book $book, EntityManagerInterface $entityManager, SluggerInterface $slugger): Response
-    {
-        if (!$this->isGranted('ROLE_ADMIN')) return $this->redirectToRoute('app_home');
-        $form = $this->createForm(BookType::class, $book);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $imageFile = $form->get('cover')->getData();
-            if ($imageFile) {
-                $originalFilename = pathinfo($imageFile->getClientOriginalName(), PATHINFO_FILENAME);
-                $safeFilename = $slugger->slug($originalFilename);
-                $newFilename = $safeFilename . '-' . uniqid() . '.' . $imageFile->guessExtension();
-                $imageFile->move( 
-                    $this->getParameter('images_directory'),
-                    $newFilename
-                ); 
-                $book->setCover($newFilename);
-            }
-
-            $entityManager->flush();
-
-            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
-        }
-
-        return $this->render('book/edit.html.twig', [
-            'book' => $book,
-            'form' => $form,
-        ]);
-    }
-}
