+++ src/Controller/DashboardController.php
@@ -13,10 +13,15 @@ final class DashboardController extends AbstractController
     #[Route('/dashboard', name: 'app_dashboard')]
     public function index(UserRepository $userRepo, BookRepository $bookRepo): Response
     {
-        if (!$this->isGranted('ROLE_ADMIN')) return $this->redirectToRoute('app_home');
-        return $this->render('dashboard/index.html.twig', [
-            'users' => $userRepo->findAll(),
-            'books' => $bookRepo->findAll()
-        ]);
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_home');
+        } else {
+            $users = $userRepo->findAll();
+            $books = $bookRepo->findAll();
+            return $this->render('dashboard/index.html.twig', [
+                'users' => $users,
+                'books' => $books
+            ]);
+        }
     }
 }
