--- src/Form/TagType.php
@@ -1,27 +1,0 @@
-<?php
-
-namespace App\Form;
-
-use App\Entity\Tag;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-class TagType extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        $builder
-        ->add('name', TextType::class, [
-            'label' => 'Catégorie :'
-        ]);
-    }
-
-    public function configureOptions(OptionsResolver $resolver): void
-    {
-        $resolver->setDefaults([
-            'data_class' => Tag::class,
-        ]);
-    }
-}
