--- templates/book/new.html.twig
@@ -1,35 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% set layout = false %}{# Disables header and footer #}
-
-{% block title %}New Book{% endblock %}
-
-{% block body %}
-    <main>
-        <div class="book-new">
-            {# <div class="btn-container"> #}
-                <a class="button-admin" href="{{ path('app_dashboard') }}"><span>Retour</span></a>
-            {# </div> #}
-            <h1>Modifier</h1>
-            {{ form_errors(form) }}
-            {{ form_start(form) }}
-            {{ form_label(form.name) }}
-            {{ form_widget(form.name) }}
-
-            {{ form_label(form.summary) }}
-            {{ form_widget(form.summary) }}
-
-            {{ form_label(form.cover) }}
-            {{ form_widget(form.cover) }}
-
-            {{ form_label(form.tags) }}
-            {{ form_widget(form.tags) }}
-
-            {# <div class="btn-container"> #}
-                <button class="button-admin"><span>Modifier</span></button>
-            {# </div> #}
-            {{ form_end(form) }}
-
-        </div>
-    </main>
-{% endblock %}
